% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_identical.R
\name{all_identical}
\alias{all_identical}
\title{Check if all list elements are identical}
\usage{
all_identical(lst)
}
\arguments{
\item{lst}{A list}
}
\value{
Logical, indicating if all elemenets are identical
}
\description{
Check if all list elements are identical
}
\examples{
all_identical(list(a=1, b=1, c=1, d=2))
}
