% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat-alt.R
\name{cat-alt}
\alias{cat-alt}
\title{testthat style cat}
\arguments{
\item{fmt}{a sprintf style format}

\item{...}{arguments matching the number of \\%s in format}
}
\value{
none
}
\description{
testthat style cat
}
\examples{
cat2("this is a '\%s' text", 'yellow')
cat3("this is a '\%s' text", 'purple')
cat4("this is a '\%s' text", 'blue')
}
