% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{char_pretty}
\alias{char_pretty}
\title{character formatted with colour}
\usage{
char_pretty(char)
}
\arguments{
\item{char}{character}
}
\value{
None
}
\description{
Can only be used with the \code{cat} function
}
\examples{
cat(sprintf("haha! this is \%s", char_pretty('cool')))
}
