% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files_exist.R
\name{check_files_exist}
\alias{check_files_exist}
\title{File checker}
\usage{
check_files_exist(input_list)
}
\arguments{
\item{file_list}{vector of full paths to files}
}
\value{
if any files are missing, will produce an error mentioning the files
}
\description{
Check input files exist
}
\author{
Al J Abadi, \email{aljalabadi@gmail.com}
}
\concept{file_recorder}
\keyword{check}
\keyword{file}
