% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_heatmap.R
\name{create_heatmap}
\alias{create_heatmap}
\title{Create heatmap}
\usage{
create_heatmap(
  mat = matrix(rnorm(100), nrow = 20),
  low = "#a4eaff",
  high = "#1C28A3",
  seg.color = "grey40",
  NA_color = "grey70",
  seg.shape = c("none", "cross", "line"),
  title = NULL,
  xtitle = NULL,
  ytitle = NULL,
  show.legend = TRUE,
  legend.title = "Expression"
)
}
\arguments{
\item{mat}{A numeric matrix with features as rows}

\item{low}{Color for lower gradient bound}

\item{high}{Color for higher gradient bound}

\item{seg.color}{Color of cell diagonal line for missing value tiles}

\item{NA_color}{Color for missing value tiles}

\item{seg.shape}{One of 'line', 'cross', or 'none'}

\item{title}{Plot title}

\item{xtitle}{X title}

\item{ytitle}{Y title}

\item{show.legend}{show.legend}

\item{legend.title}{legend.title}
}
\value{
a ggplot object
}
\description{
Creates a heatmap from a matrix, potentially with NA's
}
\examples{
create_heatmap(title = 'Example')
create_heatmap(low = 'green', high = 'purple')
create_heatmap(low = 'yellow', high = 'grey50')
create_heatmap(low = 'yellow', high = 'darkgreen')
create_heatmap(low = '#6193f2', high = '0042bf')
}
