% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirTree.R
\name{dirTree}
\alias{dirTree}
\title{create a tree of directories}
\usage{
dirTree(
  files = FALSE,
  Dir = ".",
  depth = 2,
  exclude = c("^.git"),
  plot = FALSE,
  ...
)
}
\arguments{
\item{files}{whether to look for files. If FALSE, directories are listed.}

\item{Dir}{character, the directory}

\item{depth}{the depth of the tree}

\item{exclude}{vector of regex characters to exclude}

\item{plot}{whether to plot the tree as well - not used atm}

\item{...}{Other args passed to \code{list.dirs} or \code{list.files}}
}
\description{
creates a dendogram of subdirectories down to a given depth.
}
\seealso{
Other file: 
\code{\link{check_files_exist}()},
\code{\link{file_backup}()},
\code{\link{move_symlink}()},
\code{\link{parent_base_ext}()},
\code{\link{saveRDS2}()},
\code{\link{suffixer}()}
}
\concept{file}
