% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_function_formals.R
\name{eval_function_formals}
\alias{eval_function_formals}
\title{Evaluate function formals in an environment}
\usage{
eval_function_formals(fun, envir = .GlobalEnv, ...)
}
\arguments{
\item{fun}{Name, function name}

\item{envir}{environment, in  which to evaluate. (.GlobalEnv by default)}

\item{...}{passed to base::eval}
}
\value{
Function formals are evaluated in the specified environment
}
\description{
Evaluate function formals in an environment
}
\examples{
eval_function_formals(rnorm)
}
