% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_backup.R
\name{file_backup}
\alias{file_backup}
\title{backup files - usually data files before being overwritten}
\usage{
file_backup(file, where = "/Volumes/0414861919/Mac_backup",
  date_ext = c("year", "month", "day"),
  run_spec = "Description of the latest changes you made",
  toplevel = NA)
}
\arguments{
\item{file}{The file to be recorded - relative or full path - assumes it's a full path only if it starts with '/'.}

\item{where}{Backup scheme: 1. NA: In ./backup folder of working directory.
2. 'path': Full path to the folder where a copy of the file and its log will be stored.}

\item{date_ext}{Subset of \code{c("year","month","day", "hour", "minute")} - details of the renamed file}

\item{run_spec}{Specification of the run whose output will replace the current one.}

\item{toplevel}{toplevel directory to replace with \code{where}. NA by default.}
}
\value{
backs up a renamed version of the file and logs the record.
}
\description{
Back up a file in external hard drive
}
\details{
Backs up in either 1../__backup folder or
2. in anotehr (external hard drive) folder path identical to that on local computer or
3. in another (external hard drive) folder
}
\seealso{
Other file: \code{\link{R2Excel}},
  \code{\link{check_files_exist}}, \code{\link{dirTree}},
  \code{\link{move_symlink}},
  \code{\link{parent_base_ext}}, \code{\link{saveRDS2}},
  \code{\link{suffixer}}
}
\author{
Al J Abadi, \email{aljalabadi@gmail.com}
}
\concept{file}
