% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_recorder.R
\name{file_recorder}
\alias{file_recorder}
\title{file recorder}
\usage{
file_recorder(outFile, logFolder = "__log", logFile = NULL,
  date_ext = c("year", "month", "day", "hour"), run_spec = "New run")
}
\arguments{
\item{outFile}{The file to be recorded e.g. ~/Documents/mydata.Rds - full path from the working directory}

\item{logFolder}{Full path to the record-keeping folder - full path relative to \code{outFile}. e.g. ./_log}

\item{logFile}{Name of the log file to use for records, e.g. record.log. if \code{NULL}, \code{paste0(logFolder,"/",outFile,'.log')} will be used.}

\item{date_ext}{Subset of \code{c("year","month","day", "hour", "hour", "minute")} - details of the renamed file}

\item{run_spec}{Specification of the run whose output will replace the current one.}
}
\value{
Writes a renamed version of the file to logFolder and records in logFile.
}
\description{
Create a record of the overwritten file
}
\details{
Backs up a file in \code{logFolder} before being overwriting it and logs the record
}
\examples{
ex_file <- '~/example.txt'
ex_dir <- './file_recorder_tmp'
file.create('~/example.txt')
file_recorder(outFile = ex_file, logFolder = ex_dir,
              logFile = "ex_log", run_spec = "example run with no parameter")
## remove directory/files
unlink(ex_dir)
file.remove(ex_file)
}
\author{
Al J Abadi, \email{aljalabadi@gmail.com}
}
\concept{check_files_exist}
\keyword{backup}
\keyword{file_recorder}
\keyword{log}
\keyword{overwrite}
\keyword{recorder}
