% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_version.R
\name{get_package_version}
\alias{get_package_version}
\title{Get package version from DESCRIPTION's Version line}
\usage{
get_package_version(dir = ".", branch = NULL)
}
\arguments{
\item{dir}{Character, the package directory}

\item{branch}{Character, the branch name}
}
\value{
Package's version as character x.y.z
}
\description{
Get package version from DESCRIPTION's Version line
}
\seealso{
Other gitflow: 
\code{\link{bump_up_version}()}
}
\concept{gitflow}
