% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_version.R
\name{get_package_version_and_date}
\alias{get_package_version_and_date}
\title{Get package version and date info from DESCRIPTION}
\usage{
get_package_version_and_date(dir = ".", branch = NULL)
}
\arguments{
\item{dir}{Character, the package directory}

\item{branch}{Character, the branch name}
}
\value{
Package's version as character x.y.z and date line (if any)
}
\description{
Get package version and date info from DESCRIPTION
}
\seealso{
Other git: 
\code{\link{bump_up_version}()}
}
\concept{git}
