% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_version.R
\name{get_pkg_version}
\alias{get_pkg_version}
\title{Get package version from DESCRIPTION's Version line}
\usage{
get_pkg_version(dir = ".")
}
\arguments{
\item{dir}{Package directory}
}
\value{
Package's version as character x.y.z
}
\description{
Get package version from DESCRIPTION's Version line
}
\seealso{
Other gitflow: 
\code{\link{version_bumper}()}
}
\concept{gitflow}
