% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{gg colors hue}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{Number of calolrs}
}
\value{
A character vector of \code{n} HEX colors
}
\description{
Custom color hue
}
\details{
Create\code{n} equally spaced hues around the color wheel (ggplot style)
}
\examples{
library(ggplot2)
N <- nlevels(factor(mtcars$carb))
ggplot2::ggplot(mtcars, aes(wt, disp, col=factor(carb))) + geom_point() +
  scale_color_manual(values = gg_color_hue(N))
}
\author{
Al J Abadi, \email{aljalabadi@gmail.com}
}
\keyword{hue}
