% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_mean.R
\name{gm_mean}
\alias{gm_mean}
\title{geometric mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vevtor}
}
\value{
wgeometric mean of the vector
}
\description{
Geometric Mean
}
\details{
Calculate the geometric mean of a numeric vector
}
\examples{
\dontrun{gm_mean(c(1,1,125))}
}
\seealso{
Other stats: \code{\link{wtd_var}}
}
\author{
Al J Abadi, \email{aljalabadi@gmail.com}
}
\concept{stats}
\keyword{geometric}
\keyword{mean}
