% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goEnrich.R
\name{GOget}
\alias{GOget}
\alias{goEnrich}
\alias{GOplot}
\title{GO enrichment}
\usage{
GOget(spls.obj, dat, comp, ID = "symbol", nodeSize = 10,
  mapping = "org.Mm.eg.db", subset = NULL, feas_genes = NULL)

goEnrich(GOdata, algorithm = "classic", statistic = "ks",
  topNodes = 25)

GOplot(goEnrichment, minEnrich = 0.5, trms, col,
  ymax = max(-log10(goEnrichment$KS)), minor = 0.5)
}
\arguments{
\item{spls.obj}{spls or diablo}

\item{dat}{markers <- selectVar(spls.obj, comp = comp)[\link{dat}]$value}

\item{comp}{PLS component}

\item{ID}{"symbol" or "ensembl"}

\item{nodeSize}{5-10 is good}

\item{mapping}{"org.Mm.eg.db" or "org.Hs.eg.db" package must be loaded}

\item{subset}{subset of genes, if NULL signature used}

\item{feas_genes}{feasible genes, if NULL all matrix colnames used}

\item{GOdata}{GOdata}

\item{algorithm}{passed to 'runTest``}

\item{statistic}{statistic}

\item{topNodes}{of top nodes in GO plot to keep}

\item{goEnrichment}{matrix}

\item{minEnrich}{min enrichment}

\item{trms}{terms to highlight}

\item{col}{colors for non-matching and matching terms}

\item{ymax}{max of y in the plot}

\item{minor}{minor grid}

\item{algorithm}{algorithm}

\item{topNodes}{top nodes}

\item{trms}{list of terms}

\item{col}{colors}
}
\value{
a GO term enrichment plot

data.frame of terms
}
\description{
GO terms enrichment for mixOmics output
}
