% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_values.R
\name{impute_missing_values}
\alias{impute_missing_values}
\title{Impute missing values om columns by means/median}
\usage{
impute_missing_values(x, type = c("mean", "median", "knn"), ...)
}
\arguments{
\item{x}{A numeric matrix - features are columns}

\item{type}{One of c('mean', 'median', 'knn')}

\item{...}{arguments passed to \code{impute::knn}}
}
\value{
Imputed x
}
\description{
For type = 'mean' por 'median', the corresponding values for columns
are returned. For type = 'knn', \code{impute::knn} is used.
}
\examples{
impute_by_means(matrix(c(1,2,3, NA,
                         4,5,6, NA), byrow = FALSE, ncol=2))
}
