% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_local_package.R
\name{install_local_package}
\alias{install_local_package}
\title{Install local package with dependencies}
\usage{
install_local_package(path = ".")
}
\arguments{
\item{path}{path to package}
}
\value{
none
}
\description{
Install dependencies based oon the DESCRIPTION file and then installs the
package.
}
