% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pkgs.R
\name{install_pkgs}
\alias{install_pkgs}
\title{Install a list of packages in R}
\usage{
install_pkgs(pkgs = NULL, dirpath = NULL, max_size = 1e+06)
}
\arguments{
\item{pkgs}{a character vector of packages}

\item{dirpath}{the path to a R/Rmd file containing librar() and/or require() statments, or to the /Resources/library directory
of another R version}

\item{max_size}{size (in bytes) beyond which the pkg will not be installed}
}
\value{
a data.frame containing successfull and failed/non-installation info
}
\description{
Install a list of packages in R
}
\seealso{
Other misc: \code{\link{insertInAddin}}
}
\concept{misc}
