% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pkg_version.R
\name{load_pkg_versions}
\alias{load_pkg_versions}
\title{Allow loading different versions of a package}
\usage{
load_pkg_versions(
  which = c("new", "old"),
  pkg_name = "mixOmics",
 
    new_path = "/Users/alabadi/Projects/dev/R/_work/mixOmics/mixOmics_ajabadi/mixOmics_ajabadi",
  old_path = "/Library/Frameworks/R.framework/Versions/4.0/Resources/library"
)
}
\arguments{
\item{which}{Character, on of c('new', 'old')}

\item{pkg_name}{Character, name of the package}

\item{new_path}{Character, path to new (devel) package}

\item{old_path}{Character, path to old (reference) installation}
}
\description{
Allow loading different versions of a package
}
