% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_symlink.R
\name{move_symlink}
\alias{move_symlink}
\title{file transferer}
\usage{
move_symlink(path = "/Users/alabadi/Downloads/trash",
  external = "/Volumes/0414861919/Mac_Home", home = "/Users/alabadi")
}
\arguments{
\item{path}{full path to the file/folder (~ not allowed)}
}
\value{
copies the files folder in a replicate directory in external and symlinks
}
\description{
Move a file/folder and symlink
}
\author{
Al J Abadi, \email{aljalabadi@gmail.com}
}
