% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namedVec.R
\name{namedVec}
\alias{namedVec}
\title{Create a named vector}
\usage{
namedVec(names, vec)
}
\arguments{
\item{names}{names}

\item{vec}{vector}
}
\value{
a named vector
}
\description{
Create a named vector
}
\examples{
 nemedVec(names=c('girl', 'boy'), vec=('sara', 'jack'))
}
