% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_col_pallette.R
\name{plot_col_palette}
\alias{plot_col_palette}
\title{Visualise a colour palette}
\usage{
plot_col_palette(cols)
}
\arguments{
\item{cols}{character vector of colours}
}
\value{
a plot
}
\description{
Visualise a colour palette
}
\examples{
plot_col_palette(c("#440154FF", "#482878FF", "#3E4A89FF", "#31688EFF", "#26828EFF",
"#1F9E89FF", "#35B779FF", "#6DCD59FF", "#B4DE2CFF", "#FDE725FF"
))
}
