% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2Excel.R
\name{R2Excel}
\alias{R2Excel}
\title{Data frame to excel file converter}
\usage{
R2Excel(dataf, file, sheetName, title, subtitle, overwriteSheet = TRUE)
}
\arguments{
\item{dataf}{data.frame, with rownames and column names}

\item{file}{character, relative path to the output file name}

\item{sheetName}{character, sheet name to create}

\item{sheetName}{logical, whether to replace possible existing sheet}
}
\value{
Creates a new sheet in the workbook
}
\description{
Write R outputs to Excel file
}
\examples{
R2Excel(mtcars, file='mtcars.xlsx', sheetName = 'mtcars', title = "Main title", subtitle = "More detailed description")
if(!file.exists('mtcars.xlsx'))
   stop('file was not created')
## remove
unlink('mtcars.xlsx')
}
\references{
\url{http://www.sthda.com/english/wiki/r-xlsx-package-a-quick-start-guide-to-manipulate-excel-files-in-r}
}
\seealso{
Other file: \code{\link{check_files_exist}},
  \code{\link{dirTree}}, \code{\link{file_backup}},
  \code{\link{move_symlink}},
  \code{\link{parent_base_ext}}, \code{\link{saveRDS2}},
  \code{\link{suffixer}}
}
\author{
Al J Abadi, \email{aljalabadi@gmail.com}
}
\concept{file}
\keyword{excel}
