% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_na.R
\name{rand_na}
\alias{rand_na}
\title{Random NA inseration}
\usage{
rand_na(mat = matrix(1:100, ncol = 20), prop = 0.3)
}
\arguments{
\item{mat}{a matrix, or data.frame}

\item{prop}{proportion of NAs}
}
\value{
matrix with NAs
}
\description{
Random NA inseration
}
\examples{
rand_na()
}
