% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_df_list.R
\name{rbind_df_list}
\alias{rbind_df_list}
\title{Bind lists and add element names}
\usage{
rbind_df_list(lst, new_col = "dataset")
}
\arguments{
\item{lst}{A named list containing arrays with common column names}

\item{new_col}{The names of the new column to be added as the list names}
}
\value{
A data.frame object
}
\description{
Bind lists and add element names
}
