% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveName.R
\name{saveName}
\alias{saveName}
\title{Name for output file}
\usage{
saveName(dir = "./img", basename, args, ext = "pdf", spec = NULL)
}
\arguments{
\item{dir}{directory}

\item{basename}{basename befor specs, for example: PCA-plot}

\item{args}{named arguments list that changes for different outputs}
}
\value{
a string
}
\description{
Name for output file
}
\examples{
saveName(dir='./img', basename='ggplot-of-data',
         args=list(alpha=5, beta=3), ext='pdf', spec='gamma-fixed')
}
