% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinulate_sc_genexp.R
\name{simulate_sc_genexp}
\alias{simulate_sc_genexp}
\title{Simulate matrix of single-cell gene expression counts}
\usage{
simulate_sc_genexp(
  ncells = 50,
  ngenes = 500,
  nhvgs = NULL,
  log_mu_lim = c(-1, 5),
  nsuccess = 10,
  hvgs_nsuccess = NULL
)
}
\arguments{
\item{ncells}{NUmber of cells}

\item{ngenes}{Number of genes}

\item{nhvgs}{Number of highly variable genes}

\item{log_mu_lim}{Log mean expression range}

\item{nsuccess}{Number of success in NB, proportional to the
inverse of dispersion}
}
\value{
A matrix of gene expression
}
\description{
Samples from a negative binomial distribution where log mean expression of
genes are uniformly distributed. If \code{nhvgs} is an integer, the dispersion
for \code{nhvgs} genes is increased by dividing \code{nsuccess} by 5.
}
\examples{
View(simulate_genexp())
matrixStats::rowVars(simulate_genexp()) \%>\% hist()
## 100 HVGs
matrixStats::rowVars(simulate_genexp(nhvgs = 100)) \%>\% hist()
}
