% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffixer.R
\name{suffixer}
\alias{suffixer}
\title{Add suffix to file name - even with full path}
\usage{
suffixer(file, suffix)
}
\arguments{
\item{file}{full/relative path to the file name}

\item{suffix}{descriptional suffix to add}
}
\value{
file name modified by suffix
}
\description{
Add a suffix to a file name. Could be full path.
}
\seealso{
Other file: \code{\link{check_files_exist}},
  \code{\link{dirTree}}, \code{\link{file_backup}},
  \code{\link{move_symlink}},
  \code{\link{parent_base_ext}}, \code{\link{saveRDS2}}
}
\author{
Al J Abadi, \email{aljalabadi@gmail.com}
}
\concept{file}
