% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_Altools}
\alias{update_Altools}
\title{Update the package from github}
\usage{
update_Altools(force = TRUE)
}
\arguments{
\item{force}{Logical, force re-install?}
}
\description{
Update the package from github
}
