% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{source_from_github}
\alias{source_from_github}
\title{source all R files on a github repo}
\usage{
source_from_github(repo = "ajabadi/Altools", files = NULL)
}
\arguments{
\item{repo}{of form "user/repo"}
}
\value{
sources all R files
}
\description{
source all R files on a github repo
}
