% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_bumper.R
\name{version_bumper}
\alias{version_bumper}
\title{Bump up z in x.y.z for version number}
\usage{
version_bumper(branch = "devel", dir = ".")
}
\arguments{
\item{branch}{Name of branch to bump up}

\item{dir}{Package directory}
}
\value{
Updates the version number in \code{Version: x.y.z} line by one
}
\description{
Bump up z in x.y.z for version number
}
