% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd_stats.R
\name{wtd_var}
\alias{wtd_var}
\title{weighted variance with NA's}
\usage{
wtd_var(x, weights = NULL, normwt = FALSE, na.rm = TRUE,
  method = c("unbiased", "ML"))
}
\arguments{
\item{x}{numeric vector possibly with missing values}

\item{weights}{numeric vector pof weights possibly with missing values}
}
\description{
Wrapper for Hmisc::wtd.stats
}
\details{
A detailed description
}
\references{
\url{https://github.com/harrelfe/Hmisc/blob/master/man/wtd.stats.Rd}
}
\seealso{
Other stats: \code{\link{gm_mean}}
}
\author{
Frank Harrell , Benjamin Tyner
}
\concept{stats}
