% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_funcs.R
\name{saveRDS2}
\alias{saveRDS2}
\alias{png2}
\alias{pdf2}
\title{cutom saveRDS}
\usage{
saveRDS2(object, file, force = FALSE, suffix = NULL, log = NULL, ...)

png2(file, fig_dir = "figure/", ...)

pdf2(file, fig_dir = "figure/", ...)
}
\arguments{
\item{object}{object to save}

\item{file}{full/relative path to the file name}

\item{force}{force replace?}

\item{suffix}{descriptional suffix to add}

\item{log}{a message for log file in __log directory}

\item{...}{args passed to saveRDS}

\item{figdir}{figure directory}

\item{figdir}{figure directory}
}
\value{
file name modified by suffix
}
\description{
Enhanced base saveRDs
}
\details{
Handles auto-renaming and overwrite checks. Useful when multiple files are being generated using different parameters. No need to do a lot of sprints's.
Also prevents unwanted overwrite. If suffix='auto', it'll number them.
}
\seealso{
Other file: \code{\link{check_files_exist}},
  \code{\link{dirTree}}, \code{\link{file_backup}},
  \code{\link{move_symlink}},
  \code{\link{parent_base_ext}}, \code{\link{suffixer}}
}
\author{
Al J Abadi, \email{aljalabadi@gmail.com}
}
\concept{file}
