% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bump_up_version.R
\name{bump_up_version}
\alias{bump_up_version}
\title{Bump up z in x.y.z for version number}
\usage{
bump_up_version(dir = ".", branch = "devel")
}
\arguments{
\item{dir}{Character, the package directory}

\item{branch}{Character, the name of branch to bump up}
}
\value{
Updates the version number in \code{Version: x.y.z} line by one
}
\description{
Bump up z in x.y.z for version number
}
\seealso{
Other gitflow: 
\code{\link{get_package_version}()}
}
\concept{gitflow}
