% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{char_pretty}
\alias{char_pretty}
\title{character formatted with colour}
\usage{
char_pretty(char)
}
\arguments{
\item{char}{character}
}
\value{
None
}
\description{
character formatted with colour
}
\examples{
cat(sprintf("haha! this is \%s cool", char_pretty('foo')))
}
