% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files_exist.R
\name{check_files_exist}
\alias{check_files_exist}
\title{input/ouput checker}
\usage{
check_files_exist(io_list)
}
\arguments{
\item{io_list}{list of input/output files/directories}
}
\value{
if any files/directories are missing, will produce an error mentioning the missing ones
}
\description{
check input/ouput file/directories are valid
}
\examples{
\dontrun{check_files-exist(list('~/.Rprofile'))}
}
\seealso{
Other file: 
\code{\link{dirTree}()},
\code{\link{file_backup}()},
\code{\link{move_symlink}()},
\code{\link{parent_base_ext}()},
\code{\link{saveRDS2}()},
\code{\link{suffixer}()}
}
\author{
Al J Abadi, \email{aljalabadi@gmail.com}
}
\concept{file}
\keyword{check}
\keyword{file}
