% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files_exist.R
\name{check_files_exist}
\alias{check_files_exist}
\title{File checker}
\usage{
check_files_exist(input_list)
}
\arguments{
\item{file_list}{vector of full paths to files}
}
\value{
if any files are missing, will produce an error mentioning the files
}
\description{
Check input files exist
}
\examples{
\dontrun{check_files-exist('~/.Rprofile')}
}
\seealso{
Other file: \code{\link{dirTree}},
  \code{\link{file_backup}}, \code{\link{move_symlink}},
  \code{\link{parent_base_ext}}, \code{\link{saveRDS}},
  \code{\link{suffixer}}
}
\author{
Al J Abadi, \email{aljalabadi@gmail.com}
}
\concept{file}
\keyword{check}
\keyword{file}
