% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_color_ramp.R
\name{create_color_ramp}
\alias{create_color_ramp}
\title{Create and visualise a color ramp}
\usage{
create_color_ramp(
  col_range = c("red", "yellow", "springgreen", "royalblue"),
  n = 30,
  plot = TRUE
)
}
\arguments{
\item{col_range}{Colors to transition from/to}

\item{n}{Number of colors to create}

\item{plot}{whether to plot the colors}
}
\value{
A visual of colors and a character vector of colors
}
\description{
Create and visualise a color ramp
}
\examples{
create_color_ramp(c('#a4eaff', '#0a5090'))
}
