% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirTree.R
\name{dirTree}
\alias{dirTree}
\title{create a tree of directories}
\usage{
dirTree(Dir, depth = 2, plot = FALSE)
}
\arguments{
\item{depth}{the depth of the tree}

\item{plot}{whether to plot the tree as well - not used atm}

\item{path}{the base directory}
}
\description{
creates a dendogram of subdirectories down to a given depth. Plus a tree plot.
}
