% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirTree.R
\name{dirTree}
\alias{dirTree}
\title{create a tree of directories}
\usage{
dirTree(Dir, depth = 2, plot = FALSE)
}
\arguments{
\item{depth}{the depth of the tree}

\item{plot}{whether to plot the tree as well - not used atm}

\item{path}{the base directory}
}
\description{
creates a dendogram of subdirectories down to a given depth. Plus a tree plot.
}
\seealso{
Other file: \code{\link{check_files_exist}},
  \code{\link{file_backup}}, \code{\link{move_symlink}},
  \code{\link{parent_base_ext}}, \code{\link{saveRDS}},
  \code{\link{suffixer}}
}
\concept{file}
