% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_symlink.R
\name{move_symlink}
\alias{move_symlink}
\title{file transferer}
\usage{
move_symlink(path = "/Users/alabadi/Downloads/trash",
  external = "/Volumes/0414861919/Mac_Home", home = "/Users/alabadi")
}
\arguments{
\item{path}{full path to the file/folder (~ not allowed)}

\item{external}{full path to the external directory replacing top level in 'path'}

\item{home}{the top level directory from local to be replaced by 'external' for external backup}
}
\value{
copies the files folder in a replicate directory in external and symlinks
}
\description{
Move a file/folder to external drive in similar external directory and create a symlink in local
}
\seealso{
Other file: \code{\link{check_files_exist}},
  \code{\link{dirTree}}, \code{\link{file_backup}},
  \code{\link{parent_base_ext}}
}
\author{
Al J Abadi, \email{aljalabadi@gmail.com}
}
\concept{file}
