% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_na.R
\name{rand_na}
\alias{rand_na}
\title{Random NA inseration}
\usage{
rand_na(mat = matrix(1:100, ncol = 20), prop = 0.3, seed = NULL)
}
\arguments{
\item{mat}{Matrix, or data.frame of data}

\item{prop}{Numeric b/w 0 and 1, proportion of NAs}

\item{seed}{Integer, seed number}
}
\value{
matrix with NAs
}
\description{
Random NA inseration
}
\examples{
rand_na()
}
